ISU-98 Fixes issue with "my files" not loading in media browser.
The way media renders the view that loads "my files" requires very specific
settings be loaded in the appropriate place in
Drupal.settings.ckeditor.plugins['media'].global in order to pass the
correct parameters on to the view. This configures the parameters in a such
a way that the view can find the files.