ISU-98 Add manualcrop dependency and customization.
Adds form_alters to alter the file_entity_add_upload form. Adds
descriptions to the alt and title attribute fields as they are
written in media module so that those field descriptions remain
the same across the different forms. Removes the media folder
and tags fields from the second step of the upload form so
that they aren't displayed in two different places in the
upload workflow.

Adds modification to include manualcrop in the initial file
process for image fields. On the second page of the media
browser file upload, manualcrop will display and allow the user
to crop the image during the initial upload process. This
improves the usability because it obviates the need to do a two
step process of first uploading the file, exiting the media
browser modal, and then clicking "edit" to return to a new modal
(ctools) in order to see the manualcrop options and crop the
image.