ISU-155 Change order so cancel disregards validation.
Checks if the cancel button was pressed first so that cancel
doesn't fail on validation errors. Changes the validation to
check for single character rather than two characters. The
reason for this is that normal required form validation on
drupal form fields only requires 1 character to pass. As a
result this change makes the validation consistent across
each step of the image inclusion workflow.